alterações no app

+++ src/app/pages/view-post/view-post.page.ts
@@ -8,6 +8,7 @@ import { Post } from 'src/app/models/post';
 import { PostsService } from 'src/app/services/posts.service';
 import { UserService } from 'src/app/services/user.service';
 import { Curtida } from 'src/app/models/curtidas';
+import { LoadingController } from '@ionic/angular';
 
 @Component({
   selector: 'app-view-post',
@@ -28,23 +29,37 @@ export class ViewPostPage implements OnInit {
   };
 
   user: User;
+  enableIconHeart: boolean = true;
 
   constructor(private route: ActivatedRoute,
     private postService: PostsService,
     private userService: UserService,
     private curtidaService: CurtidaService,
     private comentarioService: ComentarioService,
-    private router: Router) {
+    private router: Router,
+    public loadingController: LoadingController) {
     this.postId = this.route.snapshot.paramMap.get('id');
-    if (this.postId) {
-      postService.getPost(this.postId).then((post: Post) => {
+    
+    
+    
+  }
+
+  async ngOnInit() {
+    const loading = await this.loadingController.create({
+      message: 'carregando',
+      showBackdrop: true
+    });
+    await loading.present();
+
+    if (this.postId) {       
+      this.postService.getPost(this.postId).then(async (post: Post) => {
         this.post = post;
         this.post.id = this.postId;
-        this.post.totalComentarios = 15457;
+
         this.postService.getPostImages(post.imagens).then((result) => {
           this.postUrlImages = result;
         })
-        
+        await loading.dismiss();
         this.curtidaService.getCurtidas(this.postId).then((result) => {
           this.postCurtidas = result;        
           if (this.postCurtidas != undefined) {
@@ -55,41 +70,44 @@ export class ViewPostPage implements OnInit {
             }
           }          
         })
-
       })
     }
-  }
 
-  ngOnInit() {
     this.userService.getLogged().subscribe((user: User) => {
       this.user = user;
     })
   }
 
   clickIconHeart() {
-    if (this.user == undefined) {
-      this.router.navigate(['/login'])
-    } else {
-      if (this.curtiu == undefined) {
-        let curtida = new Curtida();
-        curtida.postId = this.postId;
-        curtida.userId = this.user.uid;
-        this.curtidaService.addCurtidas(curtida).then((result) => {
-          this.curtiu = new Curtida();
-          this.curtiu.id = result['id']
-          this.curtiu.postId = this.postId;
-          this.curtiu.userId = this.user.uid;
-          this.heartColor = 'primary'
-          this.post.totalCurtidas = this.post.totalCurtidas + 1;
-        })
+    if(this.enableIconHeart){
+      this.enableIconHeart = false;
+      if (this.user == undefined) {
+        this.router.navigate(['/login'])
       } else {
-        this.curtidaService.deleteCurtidas(this.curtiu.id).then(() => {
-          this.curtiu = undefined;
-          this.heartColor = 'dark'
-          this.post.totalCurtidas = this.post.totalCurtidas - 1;
-        })
+        if (this.curtiu == undefined) {
+          let curtida = new Curtida();
+          curtida.postId = this.postId;
+          curtida.userId = this.user.uid;
+          this.curtidaService.addCurtidas(curtida).then((result) => {
+            this.curtiu = new Curtida();
+            this.curtiu.id = result['id']
+            this.curtiu.postId = this.postId;
+            this.curtiu.userId = this.user.uid;
+            this.heartColor = 'primary'
+            this.post.totalCurtidas = this.post.totalCurtidas + 1;
+            this.enableIconHeart = true;
+          })
+        } else {
+          this.curtidaService.deleteCurtidas(this.curtiu.id).then(() => {
+            this.curtiu = undefined;
+            this.heartColor = 'dark'
+            this.post.totalCurtidas = this.post.totalCurtidas - 1;
+            this.enableIconHeart = true;
+          })
+        }
       }
     }
+    
 
   }
